added cmd parameter. defaults to dd but allows for the use of fallocate for performance. added cmd option to spec file.

+++ manifests/files.pp
@@ -17,6 +17,9 @@
 # [*timeout*]
 #   dd command exec timeout.
 #   Defaults to 300
+# [*cmd*]
+#   What command is used to create the file, dd or fallocate. dd is better tested and safer but fallocate is significantly faster.
+#   Defaults to dd
 #
 # == Examples
 #
@@ -35,20 +38,31 @@ define swap_file::files (
   $swapfilesize  = $::memorysize,
   $add_mount     = true,
   $options       = 'defaults',
-  $timeout       = 300
+  $timeout       = 300,
+  $cmd           = 'dd'
 )
 {
   # Parameter validation
   validate_re($ensure, ['^absent$', '^present$'], "Invalid ensure: ${ensure} - (Must be 'present' or 'absent')")
+  validate_re($cmd, ['^dd$', '^fallocate$'], "Invalid ensure: ${cmd} - (Must be 'dd' or 'fallocate')")
   validate_string($swapfile)
   $swapfilesize_mb = to_bytes($swapfilesize) / 1000000
   validate_bool($add_mount)
 
   if $ensure == 'present' {
-    exec { "Create swap file ${swapfile}":
-      command => "/bin/dd if=/dev/zero of=${swapfile} bs=1M count=${swapfilesize_mb}",
-      creates => $swapfile,
-      timeout => $timeout,
+
+    if $cmd == 'dd' {
+      exec { "Create swap file ${swapfile}":
+        command => "/bin/dd if=/dev/zero of=${swapfile} bs=1M count=${swapfilesize_mb}",
+        creates => $swapfile,
+        timeout => $timeout,
+      }
+    } else {
+      exec { "Create swap file ${swapfile}":
+        command => "/usr/bin/fallocate -l ${swapfilesize_mb}M ${swapfile}",
+        creates => $swapfile,
+        timeout => $timeout,
+      }
     }
     file { $swapfile:
       owner   => root,
